Cleanup AddUmbraco & AddUmbracoCore
