+++ src/Umbraco.Web.UI.NetCore/Startup.cs
@@ -32,12 +32,10 @@ namespace Umbraco.Web.UI.NetCore
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
-            var umbracoBuilder = services.AddUmbraco(_env, _config);
-            umbracoBuilder
-                .AddAllBackOfficeComponents(_env)
+            services.AddUmbraco(_env, _config)
+                .AddAllBackOfficeComponents()
                 .AddUmbracoWebsite()
                 .Build();
-
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
